Updates Login Component
